refactor(simplecas): make simplecas a package
